requirements.txt
streamlit==1.24.0
librosa==0.10.0.post2
numpy==1.23.5
scikit-learn==1.3.0
pydub==0.25.1
requests==2.31.0
openai-whisper==20231117
torch==2.1.1
ffmpeg-python==0.2.0
moviepy==1.0.3
iso639-lang>=2.1.0
langdetect>=1.0.9

